requirements.txt
requests
beautifulsoup4
pymongo
python-dateutil
streamlit
pymongo
pandas
matplotlib
seaborn
wordcloud
nltk

